Rename so readme docgen works
