+++ packages/components/src/toggle-multiple-group-control/index.ts
@@ -1,2 +1,2 @@
 export { ToggleMultipleGroupControl } from './component';
-export { ToggleMultipleGroupControlOptionIcon } from './option-icon';
+export { ToggleMultipleGroupControlOptionIcon } from './toggle-multiple-group-control-option-icon';
