+++ packages/components/src/toggle-multiple-group-control/toggle-multiple-group-control-option-icon/index.tsx
